Move remaining advanced settings widgets into new general tab
