+++ src/aiidalab_qe/app/configuration/advanced/advanced.py
@@ -12,6 +12,10 @@ from .convergence import (
     ConvergenceConfigurationSettingsModel,
     ConvergenceConfigurationSettingsPanel,
 )
+from .general import (
+    GeneralConfigurationSettingsModel,
+    GeneralConfigurationSettingsPanel,
+)
 from .hubbard import (
     HubbardConfigurationSettingsModel,
     HubbardConfigurationSettingsPanel,
@@ -37,7 +41,11 @@ class AdvancedConfigurationSettingsPanel(
     def __init__(self, model: AdvancedConfigurationSettingsModel, **kwargs):
         super().__init__(
             model=model,
-            layout={"justify_content": "space-between", **kwargs.get("layout", {})},
+            layout={
+                "justify_content": "space-between",
+                "grid_gap": "4px",
+                **kwargs.get("layout", {}),
+            },
             **kwargs,
         )
 
@@ -45,6 +53,10 @@ class AdvancedConfigurationSettingsPanel(
             self._on_spin_type_change,
             "spin_type",
         )
+
+        general_model = GeneralConfigurationSettingsModel()
+        self.general = GeneralConfigurationSettingsPanel(model=general_model)
+        model.add_model("general", general_model)
 
         convergence_model = ConvergenceConfigurationSettingsModel()
         self.convergence = ConvergenceConfigurationSettingsPanel(
@@ -71,6 +83,7 @@ class AdvancedConfigurationSettingsPanel(
         model.add_model("hubbard", hubbard_model)
 
         self.sub_settings = {
+            "general": self.general,
             "convergence": self.convergence,
             "smearing": self.smearing,
             "magnetization": self.magnetization,
@@ -90,43 +103,6 @@ class AdvancedConfigurationSettingsPanel(
         )
         self.reset_to_defaults_button.on_click(self._on_reset_to_defaults_button_click)
 
-        self.clean_workdir = ipw.Checkbox(
-            description="Delete the work directory after the calculation",
-            indent=False,
-            layout=ipw.Layout(width="fit-content", margin="5px 2px"),
-        )
-        ipw.link(
-            (self._model, "clean_workdir"),
-            (self.clean_workdir, "value"),
-        )
-
-        # Total change setting
-        self.total_charge = ipw.BoundedFloatText(
-            min=-3,
-            max=3,
-            step=0.01,
-            description="Total charge:",
-            style={"description_width": "150px"},
-        )
-        ipw.link(
-            (self._model, "total_charge"),
-            (self.total_charge, "value"),
-        )
-
-        # Van der Waals setting
-        self.van_der_waals = ipw.Dropdown(
-            description="Van der Waals correction:",
-            style={"description_width": "150px"},
-        )
-        ipw.dlink(
-            (self._model, "van_der_waals_options"),
-            (self.van_der_waals, "options"),
-        )
-        ipw.link(
-            (self._model, "van_der_waals"),
-            (self.van_der_waals, "value"),
-        )
-
         self.advanced_tabs = ipw.Tab()
         self.advanced_tabs.observe(
             self._on_advanced_tab_change,
@@ -136,16 +112,10 @@ class AdvancedConfigurationSettingsPanel(
         self.children = [
             InAppGuide(identifier="advanced-settings"),
             self.reset_to_defaults_button,
-            self.clean_workdir,
-            self.total_charge,
-            self.van_der_waals,
-            ipw.HTML("<hr>"),
             self.advanced_tabs,
         ]
 
         self.rendered = True
-
-        self.refresh()
 
         self._update_tabs()
 
@@ -159,7 +129,7 @@ class AdvancedConfigurationSettingsPanel(
         titles = []
         for identifier, model in self._model.get_models():
             subsetting = self.sub_settings[identifier]
-            if identifier == "convergence":
+            if identifier == "general":
                 subsetting.render()
             if identifier == "magnetization":
                 if self._model.spin_type != "collinear":
