+++ src/aiidalab_qe/app/configuration/advanced/general/__init__.py
@@ -1,0 +1,7 @@
+from .general import GeneralConfigurationSettingsPanel
+from .model import GeneralConfigurationSettingsModel
+
+__all__ = [
+    "GeneralConfigurationSettingsModel",
+    "GeneralConfigurationSettingsPanel",
+]
