+++ src/aiidalab_qe/app/configuration/advanced/pseudos/model.py
@@ -367,7 +367,7 @@ class PseudosConfigurationSettingsModel(
                     "PseudoDojo stringent",
                 ],
             )
-        return self._defaults.get(trait, self.traits()[trait].default_value)
+        return super()._get_default(trait)
 
     def _get_default_dictionary(self):
         return deepcopy(self._defaults["dictionary"])
