+++ src/aiidalab_qe/common/panel.py
@@ -97,10 +97,6 @@ class SettingsModel(PanelModel, HasBlockers):
     include = tl.Bool(False)
     loaded_from_process = tl.Bool(False)
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._defaults = {}
-
     def update(self):
         """Updates the model."""
         pass
@@ -134,6 +130,10 @@ class SettingsPanel(Panel[SM]):
 class ConfigurationSettingsModel(SettingsModel, Confirmable):
     """Base model for configuration settings models."""
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._defaults = {}
+
     def update(self, specific=""):
         """Updates the model.
 
@@ -143,6 +143,9 @@ class ConfigurationSettingsModel(SettingsModel, Confirmable):
             If provided, specifies the level of update.
         """
         pass
+
+    def _get_default(self, trait):
+        return self._defaults.get(trait, self.traits()[trait].default_value)
 
 
 CSM = t.TypeVar("CSM", bound=ConfigurationSettingsModel)
@@ -559,6 +562,20 @@ class ResultsModel(SettingsModel, HasProcess):
             uuid = child.uuid if child else None
         return orm.load_node(uuid) if uuid else None  # type: ignore
 
+    def save_state(self):
+        """Saves the current state of the model to the AiiDA database."""
+        node = self.fetch_process_node()
+        results = node.base.extras.get("results", {})
+        results[self.identifier] = self.get_model_state()
+        node.base.extras.set("results", results)
+
+    def load_state(self):
+        """Loads the state of the model from the AiiDA database."""
+        node = self.fetch_process_node()
+        results = node.base.extras.get("results", {})
+        if self.identifier in results:
+            self.set_model_state(results[self.identifier])
+
     def _get_child_process_status(self, which="this"):
         state, exit_message = self._get_child_state_and_exit_message(which)
         if state == "waiting":
@@ -591,20 +608,6 @@ class ResultsModel(SettingsModel, HasProcess):
             child = which if which != "this" else self.identifier
             return getattr(outputs, child) if child in outputs else AttributeDict({})
         return AttributeDict({key: getattr(node.outputs, key) for key in node.outputs})
-
-    def save_state(self):
-        """Saves the current state of the model to the AiiDA database."""
-        node = self.fetch_process_node()
-        results = node.base.extras.get("results", {})
-        results[self.identifier] = self.get_model_state()
-        node.base.extras.set("results", results)
-
-    def load_state(self):
-        """Loads the state of the model from the AiiDA database."""
-        node = self.fetch_process_node()
-        results = node.base.extras.get("results", {})
-        if self.identifier in results:
-            self.set_model_state(results[self.identifier])
 
 
 RM = t.TypeVar("RM", bound=ResultsModel)
